+++ src/php/Api/ApiFacade.php
@@ -13,6 +13,16 @@ use Phel\Api\Transfer\PhelFunction;
 final class ApiFacade extends AbstractFacade implements ApiFacadeInterface
 {
     /**
+     * @return list<string>
+     */
+    public function replComplete(string $input): array
+    {
+        return $this->getFactory()
+            ->createReplCompleter()
+            ->complete($input);
+    }
+
+    /**
      * @param list<string> $namespaces
      *
      * @return list<PhelFunction>