+++ src/php/Api/ApiFactory.php
@@ -6,8 +6,10 @@ namespace Phel\Api;
 
 use Gacela\Framework\AbstractFactory;
 use Phel\Api\Application\PhelFnNormalizer;
+use Phel\Api\Application\ReplCompleter;
 use Phel\Api\Domain\PhelFnLoaderInterface;
 use Phel\Api\Domain\PhelFnNormalizerInterface;
+use Phel\Api\Domain\ReplCompleterInterface;
 use Phel\Api\Infrastructure\PhelFnLoader;
 
 /**
@@ -15,6 +17,14 @@ use Phel\Api\Infrastructure\PhelFnLoader;
  */
 final class ApiFactory extends AbstractFactory
 {
+    public function createReplCompleter(): ReplCompleterInterface
+    {
+        return new ReplCompleter(
+            $this->createPhelFnLoader(),
+            $this->getConfig()->allNamespaces(),
+        );
+    }
+
     public function createPhelFnNormalizer(): PhelFnNormalizerInterface
     {
         return new PhelFnNormalizer(