+++ src/php/Api/Domain/PhelFnLoaderInterface.php
@@ -24,4 +24,6 @@ interface PhelFnLoaderInterface
      * }>
      */
     public function getNormalizedNativeSymbols(): array;
+
+    public function loadAllPhelFunctions(array $namespaces): void;
 }