+++ src/php/Run/RunFacadeInterface.php
@@ -25,8 +25,8 @@ interface RunFacadeInterface
     public function getAllPhelDirectories(): array;
 
     /**
-     * @param list<string> $directories
-     * @param list<string> $ns
+     * @param  list<string>  $directories
+     * @param  list<string>  $ns
      *
      * @return list<NamespaceInformation>
      */
@@ -35,7 +35,7 @@ interface RunFacadeInterface
     public function evalFile(NamespaceInformation $info): void;
 
     /**
-     * @param list<string> $paths
+     * @param  list<string>  $paths
      *
      * @return list<NamespaceInformation>
      */