+++ CHANGELOG.md
@@ -28,6 +28,7 @@ All notable changes to this project will be documented in this file.
 - Add `defexception` special form (#819)
 - Add completion to REPL (#821)
 - Add `phel->php` and `php->phel` functions (#829)
+- Add `ApiFacade::replComplete` service (#830)
 - Improve vector performance (#823)
 - Use PHPStan level 2